requirements.txt
streamlit
sounddevice
numpy
scipy
ffmpeg-python
python-dotenv
elevenlabs>=1.0.0
groq
pydub
SpeechRecognition
pyaudio==0.2.13
